ai/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
Pillow==10.4.0
psutil==6.0.0
numpy==2.1.1
onnxruntime==1.19.2
